added page align directive
